requirements.txt
Django==4.1
django-extensions==3.2.1
djangorestframework==3.14.0
django-filter==22.1